[patch] Merge `HasStorage` back into `Node` (#424)
* Refactor: merge HasStorage back into Node

* Indicate that StorageInterface is abstract

* Refactor: Move storage out of the mixin submodule

* :bug: pass node through methods

We were never getting to the cloudpickle deletion branch

* Format black

---------
